Emotional_Layer/RAG Training/Empowerapy based training/RAG_VectorDB/requirements.txt
openai>=1.30.0
faiss-cpu>=1.8.0
numpy>=1.24.0
tqdm>=4.66.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
